Refactor network configuration handling and introduce Zustand store for state management
